requirements.txt
tensorflow==2.0.0
torch
Pillow
numpy
opencv_contrib_python
scipy
tqdm